+++ CHANGELOG.md
@@ -11,6 +11,9 @@ SPDX-License-Identifier: EUPL-1.2
 - `Machine.block` command to directly control block temperatures.
 - Certificate support for SSL, and new connection recommendations.
 - Stage lines and event spans now are now not plotted if they are outside of the selected time range / stages.
+- `Step` now supports repeats.
+- `Stage.stepped_ramp` can now take multiple data points per temperature.
+- Fix hanging connection bug.
 
 ## Version 0.11.0